Resolve #2326: tighten CLI public lazy facades
